requirements.txt
langchain==0.3.27
langchain-core==0.3.72
langchain-community>=0.0.22
langchain-google-genai>=0.0.10
google-search-results==2.4.2
wikipedia==1.4.0
python-dotenv==1.1.1
pypdf>=3.17.0
chromadb>=0.4.24
